chore: update files, move methods arround
