+++ lesson_13/bank/bank_app/src/main/java/com/codedifferently/lesson13/bank/BankAccounts.java
@@ -1,7 +1,9 @@
 package com.codedifferently.lesson13.bank;
 
-import com.codedifferently.lesson13.bank.exceptions.InsufficientFundsException;
 import java.util.Set;
+
+import com.codedifferently.lesson13.bank.exceptions.InsufficientFundsException;
+
 
 public interface BankAccounts {
   /**
