+++ lesson_13/bank/bank_app/src/test/java/com/codedifferently/lesson13/bank/BankAtmTest.java
@@ -1,14 +1,15 @@
 package com.codedifferently.lesson13.bank;
 
+import java.util.Set;
+import java.util.UUID;
+
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatExceptionOfType;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
 
 import com.codedifferently.lesson13.bank.exceptions.AccountNotFoundException;
 import com.codedifferently.lesson13.bank.exceptions.CheckVoidedException;
-import java.util.Set;
-import java.util.UUID;
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.Test;
 
 class BankAtmTest {
